fix(storage): fixup blob signing
The previous implementation via the iamcredentials API seems to no
longer be working. Signing the blob via standard PEM->PKCS methods seems
to work just fine, fortunately!

Basically, rather than offloading the signing work, we just do this
locally, so this should be a performance improvement as well. Since our
previous implementation only worked for SERVICE_ACCOUNT tokens *anyway*
(due to a mistake -- theoretically, we could have pulled the email from
the metadata server for GCE_METADATA tokens...), this shouldn't break
backwards compatibility.

Note that at the same time as the iamcredentials API breaking, it seems
the canonical uri for blob signs also was updated (the bucket name is
now included in the host instead of as part of the path). I suspect the
two are related.