fix(storage): upload retry no longer results in IO error
The retry mechanism in storage._do_upload() was broken because when you
perform a request on a stream in aiohttp, the stream is automatically
closed. Therefore, the retry raises an error: I/O Operation on closed
file.

This fix monkeypatch the close function on the stream to ensure
it remains open until all retries are performed.