chore(lint): enable mccabe complexity plugin with default settings
Enable cyclomatic complexity plugin, which uses the McCabe Cyclomatic
complexity to define how complex a function is. By default the maximum
cyclomatic complexity is `10`.

[Cyclomatic complexity](https://en.wikipedia.org/wiki/Cyclomatic_complexity)
is the number of linearly independent paths within it, or in layman's terms:
how many paths can be taken "through the code". A cyclomatic complexity of `10`
means there's 10 linearly independent paths to exit a function.

And a bonus: pin importlib<5.0 in order to fix a recent flake8
dependency issue.