Use aiohttp with contextmanager
I'm getting lots of errors like this in my logs: `Unclosed client session_client_session: <aiogoogle.sessions.aiohttp_session.AiohttpSession object at 0x7f1cb8188370>`

I believe this is because the aiohttp library is not used correctly. According to the docs, all get/post/.. request methods should always be called with a contextmanager. 

This is just a quick&dirty PR to demonstrate what I mean. Happy to provide a proper PR that changes this for all usages, if you agree with the change.