chore(pubsub): clean up log messages
Main goal: ensure that `hash(message)` is useful: eg. that log messages
can be relied on to be constant for cases where it is "the same error"
and different for cases where it is not. This allows for clients to do
things such as alert when the rate of a given error message is above a
given threshold.

While I was here, I cleaned up the other log strings in the file to make
'em all consistently typed/formatted/etc.