feat(clients): allow non-dev API endpoint roots
Specifying an API endpoint root to a client used to imply that the
endpoint in use was for dev, which would disable TLS and token bearer
authorisation. This is not always a valid assumption, for example when
manually specifying a locational endpoint for Google PubSub to target a
specific region, as such endpoints are for production and should
therefore use TLS and authorisation.

Fix this by allowing manual configuration of the `api_is_dev` setting
when using a non-dev root, whilst maintaining the old behaviour by
default for backwards compatibility.