feat(pubsub): allow explicit ack/nack on messages
Adds new ack() and nack() methods on SubscriberMessage which allow users
to forcibly mark a messsage as getting acked or nacked, regardless of
the success of the callback. In plain English:

When `callback` terminates, if `.ack()` has been called it acks. Else,
if `.nack()` has been called it nacks. Else, it acks if no exception was
raised and nacks otherwise.

Fixes #837